pyproject.toml
[tool.poetry]
name = "parse"
version = "1.0.0"
description = "sanic"
authors = ["r.zabbarov <gdeon1337@gmail.com>", "v.maleev <chestny73@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
dataclasses = {version = "^0.6",python = "<3.7"}
six = "^1.13"
toml = "^0.10.0"
sanic-envconfig = "^1.3"
pyroute2 = "^0.5.7"
sanic = "^19.6 <19.9"
sanic-openapi = {git = "https://github.com/SandLabs/sanic-openapi.git",branch = "master"}
sanic-jwt = "^1.3"
argon2-cffi = "^19.2"
aioredis = "^1.3"
tenacity = "^6.0.0"
fake-useragent = "^0.1.11"
gino = "^0.8.5"
sanic-cors = "^0.9.9"
msgpack = "^1.0.0"
ddtrace = "^0.38.2"

[tool.poetry.dev-dependencies]
prospector = "^1.1"
pytest = "^5.2"
pytest-sanic = "^1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
